move usecase to folder
