+++ apps/smtp/src/pages/api/webhooks/gift-card-sent.ts
@@ -4,7 +4,7 @@ import { saleorApp } from "../../../saleor-app";
 import { GiftCardSentWebhookPayloadFragment } from "../../../../generated/graphql";
 import { withOtel } from "@saleor/apps-otel";
 import { createLogger } from "../../../logger";
-import { SendEventMessagesUseCaseFactory } from "../../../modules/event-handlers/send-event-messages.use-case.factory";
+import { SendEventMessagesUseCaseFactory } from "../../../modules/event-handlers/use-case/send-event-messages.use-case.factory";
 
 const GiftCardSentWebhookPayload = gql`
   fragment GiftCardSentWebhookPayload on GiftCardSent {
