+++ apps/smtp/src/pages/api/webhooks/notify.ts
@@ -3,7 +3,7 @@ import { saleorApp } from "../../../saleor-app";
 import { notifyEventMapping, NotifySubscriptionPayload } from "../../../lib/notify-event-types";
 import { withOtel } from "@saleor/apps-otel";
 import { createLogger } from "../../../logger";
-import { SendEventMessagesUseCaseFactory } from "../../../modules/event-handlers/send-event-messages.use-case.factory";
+import { SendEventMessagesUseCaseFactory } from "../../../modules/event-handlers/use-case/send-event-messages.use-case.factory";
 
 /*
  * The Notify webhook is triggered on multiple Saleor events.
