+++ apps/smtp/src/pages/api/webhooks/order-created.ts
@@ -5,7 +5,7 @@ import { saleorApp } from "../../../saleor-app";
 import { OrderCreatedWebhookPayloadFragment } from "../../../../generated/graphql";
 import { withOtel } from "@saleor/apps-otel";
 import { createLogger } from "../../../logger";
-import { SendEventMessagesUseCaseFactory } from "../../../modules/event-handlers/send-event-messages.use-case.factory";
+import { SendEventMessagesUseCaseFactory } from "../../../modules/event-handlers/use-case/send-event-messages.use-case.factory";
 
 const OrderCreatedWebhookPayload = gql`
   ${OrderDetailsFragmentDoc}
