+++ apps/smtp/src/pages/api/webhooks/order-refunded.ts
@@ -7,7 +7,7 @@ import { gql } from "urql";
 import { saleorApp } from "../../../saleor-app";
 import { withOtel } from "@saleor/apps-otel";
 import { createLogger } from "../../../logger";
-import { SendEventMessagesUseCaseFactory } from "../../../modules/event-handlers/send-event-messages.use-case.factory";
+import { SendEventMessagesUseCaseFactory } from "../../../modules/event-handlers/use-case/send-event-messages.use-case.factory";
 
 const OrderRefundedWebhookPayload = gql`
   ${OrderDetailsFragmentDoc}
